refactor: move env.js out of helpers
